♻️ move non-common elements to operations modal
