+++ components/OperationsModal/ModalTxCost/index.tsx
@@ -1,11 +1,11 @@
 import React, { useMemo } from 'react';
 import { BigNumber, formatFixed } from '@ethersproject/bignumber';
 import { Typography, Skeleton, Box } from '@mui/material';
+import { useTranslation } from 'react-i18next';
 
 import { WeiPerEther } from '@ethersproject/constants';
-import ModalInfo from '../ModalInfo';
+import ModalInfo from 'components/common/modal/ModalInfo';
 import useAccountData from 'hooks/useAccountData';
-import { useTranslation } from 'react-i18next';
 
 type Props = {
   gasCost?: BigNumber;
